added correct line for routing
The onInit method contains the wrong reference for routing. It is using UIComponent instead of this.getRouter. This results currently in a TS error, as UIComponent is not known. Looking at the example source code ex6/com.myorg.myapp/src/controller/IncidenceDetail.controller.ts the correct code is used:
this.getRouter().getRoute...

I added the line so the example and the source code match.